+++ src/clojure_lsp/feature/rename.clj
@@ -107,7 +107,9 @@
   [definition references source-paths]
   (and (seq references)
        (or (not (seq source-paths))
-         (some #(string/starts-with? (:filename definition) %) source-paths))))
+           (some #(string/starts-with? (:filename definition) %) source-paths))
+       (not (and (= :keywords (:bucket definition))
+                 (not (:ns definition))))))
 
 (defn rename
   [uri new-name row col]